CLiCkin2DaBeaT: Fix app crash issue
This application has the component name for the browser hardcoded,
which causes the application to crash on systems that use GMS browser.
This hardcoded component name matches the AOSP browser, but it does not match
GMS browser. When having GMS browser installed, it will not launch
because the component name in the application does not match the
system package name.

In order to be compatible with all browser versions, we remove the
hardcoded component name.

Signed-off-by: Hua Wang <Hua.Wang@windriver.com>
Signed-off-by: Kwanhong Lee <kwanhong.lee@windriver.com>
Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>