Set minimum target version and target version
Set minimum target version and target version in Android Manifest.

For mininum target at least SDK 4 is needed (to allow support for
targetSdkVersion). In other applications, the minimum target was
chosen so that the application will compile using the respective SDK.

Target version was chosen the latest current SDK version (16).

Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>