Target the app to an older SDK revision
In the apps-for-android apps, we have seen an issue where no menu button is
display on ICS.

Per this blog post:
http://android-developers.blogspot.com/2012/01/say-goodbye-to-menu-button.html

ICS is making the assumption that if the app is targeted at ICS, it is using
the features of ICS correctly. Our version of this app, compiled in our
source tree, and being implicitly targeted at SDK version 14 (because it is
not specified in the apps manifest).

ICS is making a decision about showing or not the "legacy overflow" button
based on this targeted SDK version, so not displaying the menu button.

Signed-off-by: Ken Lierman <ken.lierman@windriver.com>
Signed-off-by: Kwanhong Lee <kwanhong.lee@windriver.com>
Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>