Photostream: fix mime type for cropped image
Gallery's crop implementation tries to determine the
compress format for the output image. Photostream
saves this image under the name "wallpaper", so there
is no extension. When Gallery will try to save the
image it will crash because it will find mime type NULL.

This fix changes the name for the output image to
"wallpaper.jpg" so that the format will be jpg
instead of unknown.

Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>