Photostream set wallpaper fix
Fixed a bug that prevented the ability to set wallpapers with the
Photostream app.
One problem was sending a malformed URI to the wallpaper cropping
application (it was missing a '/'). The second problem was that
the wallpaper being downloaded was being overriden with an empty
file when an output stream was opened for it by the cropping app.
To fix this, I've modified the Photostream app to use a different
file for storing the cropped wallpaper.

Issue: APDEV-1425
Category: AOSP Enhancement
Domain: AOSP-apps
Origin: internal
Upstream-Candidate: no

Change-Id: I783c24529d85d3f4a830dd01ef99a730265aebb2
Signed-off-by: Tiberiu Breana <tiberiu.a.breana@intel.com>
Reviewed-on: https://android.intel.com/163008
Reviewed-by: Chiras, Robert <robert.chiras@intel.com>
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Fagerstedt, Axel <axel.fagerstedt@intel.com>
Reviewed-by: Grape, Markus <markus.grape@intel.com>
Reviewed-by: Timohi, Alexandru <alexandru.timohi@intel.com>
Reviewed-by: Tomaszewski, Piotr <piotr.tomaszewski@intel.com>